refactor: extract useModelVisibilityHandlers (#3501 Phase 1l)
